+++ HeeHeeWeak/Classes/Main/Controllers/HotActivityViewController.m
@@ -116,7 +116,7 @@
     //获取网络数据
     AFHTTPSessionManager *manger = [AFHTTPSessionManager manager];
     manger.responseSerializer.acceptableContentTypes = [NSSet setWithObjects:@"text/html", nil];
-    [manger GET:[NSString stringWithFormat:@"%@&page=%ld",kHotActivityThem,_pageCount] parameters:nil progress:^(NSProgress * _Nonnull downloadProgress) {
+    [manger GET:[NSString stringWithFormat:@"%@&page=%ld",kHotActivityThem,(long)_pageCount] parameters:nil progress:^(NSProgress * _Nonnull downloadProgress) {
         
     } success:^(NSURLSessionDataTask * _Nonnull task, id  _Nullable responseObject) {
         NSDictionary *dict = responseObject;
